Revert "move generated sprite file to public folder"
This reverts commit 5da7a45cc200e3b242f26b91b1dd71e4ecea7271.
