+++ ember-flight-icons/addon/components/flight-icon-sprite.js
